rename service to server
